+++ ChangeLog.md
@@ -9,6 +9,7 @@
     - Change default positioning from 'relative' to 'absolute'
     - Let the application configure the button classes (todo #1, todo #2)
     - Let the application configure the button shape (todo #3)
+    - Let the application configure the button title (todo #4)
 
 ### 0.10.1